requirements.txt
python-owasp-zap-v2.4>=0.0.20
requests>=2.28.1
pyyaml>=6.0
beautifulsoup4>=4.11.1
lxml>=4.9.1

openai>=1.0.0
langchain>=0.0.267
anthropic>=0.2.0

python-dotenv>=0.21.0
rich>=12.6.0
tqdm>=4.64.1

pytest>=7.3.1

bandit>=1.7.5
semgrep>=1.34.0

esprima>=4.0.1
jsbeautifier>=1.14.7